requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
